move RequestAuthorizeSignatureV2 class to java file.
